backend/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
djoser==2.2.0
psycopg2-binary==2.9.3
Pillow==10.0.0
django-debug-toolbar==3.2.4
django-filter==23.2